Move send_message to celery task * add transaction.atomic
